+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/DependencyManagementPluginKitTest.groovy
@@ -12,7 +12,7 @@ class DependencyManagementPluginKitTest extends AbstractKitTest {
             plugins {
                 id 'java'
                 id 'ru.vyarus.pom'
-                id "io.spring.dependency-management" version "1.0.3.RELEASE"
+                id "io.spring.dependency-management" version "1.0.11.RELEASE"
             }
 
             group 'ru.vyarus'
@@ -65,7 +65,7 @@ class DependencyManagementPluginKitTest extends AbstractKitTest {
             plugins {
                 id 'java'
                 id 'ru.vyarus.pom'
-                id "io.spring.dependency-management" version "1.0.3.RELEASE"
+                id "io.spring.dependency-management" version "1.0.11.RELEASE"
             }
 
             group 'ru.vyarus'
@@ -126,7 +126,7 @@ class DependencyManagementPluginKitTest extends AbstractKitTest {
             plugins {
                 id 'java'
                 id 'ru.vyarus.pom'
-                id "io.spring.dependency-management" version "1.0.8.RELEASE"
+                id "io.spring.dependency-management" version "1.0.11.RELEASE"
             }
 
             group 'ru.vyarus'